@@ -1,26 +1,29 @@
 {
   "name": "ffxiv-act-chat-extractor",
-  "version": "0.1.0",
+  "version": "0.1.2",
   "private": true,
   "dependencies": {
-    "@emotion/core": "^10.0.28",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^24.0.0",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
-    "@types/react-toggle": "^4.0.2",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/material": "^5.9.0",
     "customize-cra": "^1.0.0",
-    "framer-motion": "^2.9.4",
-    "react": "^16.13.1",
-    "react-app-rewired": "^2.1.6",
-    "react-dom": "^16.13.1",
-    "react-dropzone": "^11.0.2",
-    "react-scripts": "3.4.1",
-    "react-toggle": "^4.1.1",
-    "typescript": "~3.7.2"
+    "framer-motion": "^6.4.3",
+    "react": "^18.2.0",
+    "react-app-rewired": "^2.2.1",
+    "react-dom": "^18.2.0",
+    "react-dropzone": "^14.2.2",
+    "react-scripts": "^5.0.1"
+  },
+  "devDependencies": {
+    "@emotion/babel-preset-css-prop": "^11.2.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.1",
+    "@types/jest": "^28.1.4",
+    "@types/node": "^18.0.3",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "typescript": "^4.7.4"
   },
   "scripts": {
     "start": "react-app-rewired start",
@@ -42,8 +45,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@emotion/babel-preset-css-prop": "^10.0.27"
   }
 }
