@@ -1,3 +1,5 @@
-const { override, addBabelPreset } = require("customize-cra");
+const { override, addBabelPreset } = require('customize-cra');
 
-module.exports = override(addBabelPreset("@emotion/babel-preset-css-prop"));
+module.exports = override(
+  addBabelPreset('@emotion/babel-preset-css-prop'),
+);
