@@ -1,21 +1,21 @@
 export enum CHANNEL_CODES {
-  SAY = '000a',
-  SHOUT = '000b',
-  PARTY = '000e',
-  ALLIANCE = '000f',
-  LINKSHELL1 = '0010',
-  LINKSHELL2 = '0011',
-  LINKSHELL3 = '0012',
-  LINKSHELL4 = '0013',
-  LINKSHELL5 = '0014',
-  LINKSHELL6 = '0015',
-  LINKSHELL7 = '0016',
-  LINKSHELL8 = '0017',
-  FREE_COMPANY = '0018',
-  CUSTOM_EMOTE = '001c',
-  EMOTE = '001d',
-  YELL = '001e',
-  WHISPER = '0039',
+  SAY                   = '000A',
+  SHOUT                 = '000B',
+  PARTY                 = '000E',
+  ALLIANCE              = '000F',
+  LINKSHELL1            = '0010',
+  LINKSHELL2            = '0011',
+  LINKSHELL3            = '0012',
+  LINKSHELL4            = '0013',
+  LINKSHELL5            = '0014',
+  LINKSHELL6            = '0015',
+  LINKSHELL7            = '0016',
+  LINKSHELL8            = '0017',
+  FREE_COMPANY          = '0018',
+  CUSTOM_EMOTE          = '001C',
+  EMOTE                 = '001D',
+  YELL                  = '001E',
+  WHISPER               = '0039',
   CROSSWORLD_LINKSHELL1 = '0064',
   CROSSWORLD_LINKSHELL2 = '0065',
   CROSSWORLD_LINKSHELL3 = '0066',
@@ -23,17 +23,18 @@ export enum CHANNEL_CODES {
   CROSSWORLD_LINKSHELL5 = '0068',
 }
 
+
 export const CHANNELS = [
   // Technically not a channel, but used for optional filtering
   { code: 'time', name: 'Show Timestamp' },
-  { code: '000a', name: 'Say', color: '#FFFFFF' },
-  { code: '000b', name: 'Shout', color: '#EB9234' },
-  { code: '000e', name: 'Party', color: '#1ECCE3' },
-  { code: '000f', name: 'Alliance', color: '#EB9234' },
+  { code: '000A', name: 'Say', color: '#FFFFFF' },
+  { code: '000B', name: 'Shout', color: '#EB9234' },
+  { code: '000E', name: 'Party', color: '#1ECCE3' },
+  { code: '000F', name: 'Alliance', color: '#EB9234' },
   { code: '0018', name: 'Free Company', color: '#54B068' },
-  { code: '001c', name: 'Custom Emote', color: '#8CCCC3' },
-  { code: '001d', name: 'Emote', color: '#FFFFFF' },
-  { code: '001e', name: 'Yell', color: '#C9C722' },
+  { code: '001C', name: 'Custom Emote', color: '#8CCCC3' },
+  { code: '001D', name: 'Emote', color: '#FFFFFF' },
+  { code: '001E', name: 'Yell', color: '#C9C722' },
   { code: '0039', name: 'Whisper', color: '#D439CC' },
   { code: '0010', name: 'Linkshell 1', color: '#31B56A' },
   { code: '0011', name: 'Linkshell 2', color: '#2A8C9C' },
